# service/types.go
package service

import (
	"ginchat/models"
)

type JSONResult struct {
	Code    int         `json:"code"`
	Message string      `json:"message"`
	Data    interface{} `json:"data"`
}

type ListResp struct {
	Code    int         `json:"code"`
	Message string      `json:"message"`
	Data    interface{} `json:"data"`
	Rows    interface{}
	Total   int
}

type PhoneS struct {
	Phone string `json:"phone" valid:"mobile"`
}

type EmailS struct {
	Email string `json:"email" valid:"email"`
}

type UserRegisterReq struct {
	UserName   string `form:"name" json:"userName" binding:"required"`     //	用户名
	Password   string `form:"password" json:"password" binding:"required"` //	密码
	RePassword string `form:"Identity" json:"Identity" binding:"required"` //	密码确认
}

type UserRegisterResp struct {
	models.UserBasic
}

type UserDelReq struct {
	ID uint `json:"id" binding:"required"`
}

type UserGetReq struct {
	ID uint `json:"id" binding:"required"`
}

type UserUpdateReq struct {
	ID       uint   `json:"id" binding:"required"`
	Name     string `json:"name"`
	PassWord string `json:"password"`
	PhoneS
	EmailS
}

type UserLoginReq struct {
	Name     string `json:"name" binding:"required"`
	PassWord string `json:"password" binding:"required"`
}

type UserLoginResp struct {
	Token        string `json:"token"`
	AccessExpire string `json:"accessExpire"`
	RefreshAfter string `json:"refreshAfter"`
}

type FindUserReq struct {
	Name     string `form:"name" binding:"required"`
	PassWord string `form:"password" binding:"required"`
}

type FindUserResp struct {
	models.UserBasic
}

type ToChatReq struct {
	UserId uint   `form:"userId"`
	Token  string `form:"token"`
}

type SearchFriendReq struct {
	UserId uint `form:"userId"`
}

type SearchFriendResp struct {
	JSONResult
	Rows  []models.UserBasic
	Total int
}

type ChatReq struct {
	UserId int64  `json:"userId" form:"userId"`
	Token  string `json:"token" form:"token"`
}